Move PdbReader to the Pdb namespace
All Pdb related functionality should be in the Pdb namespace.
